+++ SwoopMarketplaceProjectFrontend/Dtos/ReportDto.cs
@@ -2,11 +2,11 @@
 {
     public class ReportDto
     {
-        public int ReportId { get; set; }
+        public long ReportId { get; set; }
 
-        public int ListingId { get; set; }
+        public long ListingId { get; set; }
 
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         public string Description { get; set; } = null!;
     }
